refactor(agent-auth): split harness-settings service out of the gateway service (max-lines)
The harness_settings rider work pushed agent_gateway/service.py to 827
lines, past its 813-line allowlist. Settings are explicitly not auth and
the store layer already draws this seam
(db/store/agent_gateway/harness_settings.py), so put_harness_settings
moves verbatim to a matching service module,
server/cloud/agent_gateway/harness_settings.py, and api.py calls it
there. service.py lands at 799 lines — back under the 800 server-layer
cap — so its allowlist entry is REMOVED (the checker fails stale entries;
this deletes an exception rather than adding or raising one). No checker
or coverage change; no wire-contract change, so the generated SDK is
untouched.
